+++ lib/pages/home.dart
@@ -1,0 +1,40 @@
+import 'package:easy_localization/easy_localization.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:styled_widget/styled_widget.dart';
+
+import '../providers/counter.dart';
+
+class HomePage extends ConsumerWidget {
+  const HomePage({super.key});
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    context.locale; // force rebuild when locale changes
+
+    final counter = ref.watch(counterProvider);
+
+    // This method is rerun every time setState is called, for instance as done
+    // by the _incrementCounter method above.
+    //
+    // The Flutter framework has been optimized to make rerunning build methods
+    // fast, so that you can just rebuild anything that needs updating rather
+    // than having to individually change instances of widgets.
+    return Scaffold(
+      appBar: AppBar(
+        // Here we take the value from the MyHomePage object that was created by
+        // the App.build method, and use it to set our appbar title.
+        title: Text('home.title'.tr()),
+      ),
+      body: <Widget>[
+        Text('home.counter_text'.tr()),
+        Text('$counter').fontSize(32),
+      ].toColumn(mainAxisAlignment: MainAxisAlignment.center).center(),
+      floatingActionButton: FloatingActionButton(
+        onPressed: () => ref.read(counterProvider.notifier).increment(),
+        tooltip: 'home.increment'.tr(),
+        child: const Icon(Icons.add),
+      ), // This trailing comma makes auto-formatting nicer for build methods.
+    );
+  }
+}
