+++ lib/providers/android_info.dart
@@ -1,0 +1,5 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+import '../utilities/android_info.dart';
+
+final androidInfoProvider = FutureProvider((ref) => getAndroidInfo());
